+++ src/renderer/components/configuration-wizard/Wizard.tsx
@@ -21,7 +21,7 @@ import InstallationType from '../stages/installation/InstallTypeSelection';
 import { selectLoading } from './wizardSlice';
 import { useAppSelector } from '../../hooks';
 import InitApfAuth from '../stages/InitApfAuth';
-import { initProgress } from '../stages/progress/progressStore';
+import { initProgress } from '../stages/progress/progressStatus';
 
 export const stages = [
   {id: 0, label: 'Connection', component: <Connection/>, hasJCL: false, isSkippable: false, isSkipped: false, hasOutput: false, steps: 1, nextButton: 'Continue'},
