+++ src/renderer/components/stages/InitApfAuth.tsx
@@ -23,7 +23,7 @@ import { selectInstallationArgs } from "./installation/installationSlice";
 import { createTheme } from '@mui/material/styles';
 import { stages } from "../configuration-wizard/Wizard";
 import { setActiveStep } from "./progress/activeStepSlice";
-import { getStageDetails, getSubStageDetails } from "./progress/progressStore"; 
+import { getStageDetails, getSubStageDetails } from "../../../utils/stageDetails"; 
 
 const InitApfAuth = () => {
